requirements.txt
# Neo4j Database Driver
neo4j==5.16.0

# Anthropic Claude API Client
anthropic==0.39.0

# Environment Variable Management
python-dotenv==1.0.0

# Data Validation and Settings Management
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
requests==2.31.0

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and Utilities
colorlog==6.8.0

# Date/Time Handling
python-dateutil==2.8.2

# Type Checking (Development)
mypy==1.7.1
types-requests==2.31.0.10

# Code Quality (Development)
black==23.12.1
flake8==7.0.0
isort==5.13.2

# JSON Schema Validation
jsonschema==4.20.0
